fix: login modal closing issue and dashboard redirection
