+++ src/components/common/UserDashboardSidebar.tsx
@@ -1,14 +1,17 @@
 "use client";
 import React from "react";
-import LogoutModal from "@/src/components/common/LogoutModal";
+import LogoutModal from "@/src/components/modals/LogoutModal";
 import { LogOut } from "lucide-react";
 import { cn } from "@/src/lib/utils";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { UserDashboardSidebarItems } from "@/src/constants/UserDashboardSidebarItems";
+import LoginModal from "../modals/LoginModal";
+import { Button } from "../ui/button";
 
 const UserDashboardSidebar = () => {
   const pathname = usePathname();
+  const user = false;
   return (
     <aside className="flex w-80 flex-col rounded-xl border-r border-gray-800 bg-[#111] p-4">
       <div className="flex items-center gap-3 p-3">
@@ -17,6 +20,14 @@ const UserDashboardSidebar = () => {
           <p className="font-semibold">Siyam Ahmed</p>
           <p className="text-xs text-gray-400">UID: 496496</p>
         </div>
+
+        <LoginModal
+          TriggerButton={
+            <Button variant="outline" size="sm" className="text-xs">
+              {user ? "Log out" : "Log in"}
+            </Button>
+          }
+        />
       </div>
 
       <nav className="mt-6 flex flex-col gap-2">
