+++ src/components/modals/LoginModal.tsx
@@ -3,13 +3,11 @@
 import React, { useState } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Button } from "@/src/components/ui/button";
 import { LoginFormData, loginSchema } from "@/src/schema/loginSchema.schema";
-import { FormInputField } from "@/src/components/forms/FormInputField";
+import { LoginFormCard } from "@/src/components/modals/LoginFormCard";
 import {
   Dialog,
   DialogContent,
-  DialogDescription,
   DialogHeader,
   DialogTitle,
   DialogTrigger,
@@ -26,31 +24,12 @@ export default function LoginModal({TriggerButton}:LoginModalProps) {
   const user = false;
   const [isOpen, setIsOpen] = useState(false);
 
-  const {
-    handleSubmit,
-    formState: { isSubmitting },
-    control,
-  } = useForm<LoginFormData>({
-    resolver: zodResolver(loginSchema),
-    mode: "onChange",
-  });
-
-  // ✅ Submit handler
-  const onSubmit = (data: LoginFormData) => {
-    console.log("Login Data:", data);
-    // 🔐 here you will call API for user login
-  };
-
   return (
     <Dialog open={isOpen} onOpenChange={setIsOpen}>
       <DialogTrigger asChild>
         {TriggerButton}
       </DialogTrigger>
-      <DialogContent 
-        onInteractOutside={(e) => {
-          e.preventDefault();
-        }}
-        className="min-w-[521px] rounded-2xl border-none bg-[#0B0000] p-9">
+      <DialogContent className="min-w-[521px] rounded-2xl border-none bg-[#0B0000] p-9">
         <DialogHeader className="gap-3">
           <DialogTitle className="text-center text-2xl font-bold text-white">Login</DialogTitle>
         </DialogHeader>
@@ -104,52 +83,7 @@ export default function LoginModal({TriggerButton}:LoginModalProps) {
           </div>
         </div>
 
-        {/* Login Card */}
-          <div className="w-full max-w-md rounded-lg shadow-lg">
-            <form onSubmit={(e) => {
-              e.preventDefault();
-              setIsOpen(false);
-              handleSubmit(onSubmit)
-            }} className="space-y-4">
-              {/* Email */}
-              <FormInputField
-                name="email"
-                control={control}
-                placeholder="Enter Address"
-                type="email"
-                colorScheme={{
-                  background: "bg-[#16151A]",
-                  text: "text-white",
-                  placeholder: "placeholder-[#A9A9A9]",
-                  border: "border-[#595756]",
-                }}
-              />
-
-              {/* Password */}
-              <FormInputField
-                name="password"
-                control={control}
-                placeholder="Password"
-                type="password"
-                as="input"
-                colorScheme={{
-                  background: "bg-[#16151A]",
-                  text: "text-white",
-                  placeholder: "placeholder-[#A9A9A9]",
-                  border: "border-[#595756]",
-                }}
-              />
-              <p className="text-primary-rose cursor-pointer text-right">Forget Password?</p>
-              {/* Submit Button */}
-              <Button
-                type="submit"
-                className="bg-primary-rose hover:bg-primary-rose-hover w-full font-semibold"
-                disabled={isSubmitting}
-              >
-                {isSubmitting ? "Logging in..." : "Login"}
-              </Button>
-            </form>
-          </div>
+        <LoginFormCard onClose={() => setIsOpen(false)} />
       </DialogContent>
     </Dialog>
   );
