+++ src/components/pages/home/UserButton.tsx
@@ -5,14 +5,14 @@ import { Avatar, AvatarFallback, AvatarImage } from "../../ui/avatar";
 import { User } from "lucide-react";
 import { Separator } from "../../ui/separator";
 import Link from "next/link";
-import LoginModal from "../../modals/Login";
 
 const UserButton = () => {
   const user = false;
   return (
-    <Link href={"/dashboard"} className="flex items-center space-x-4">
+    <div>
       <HoverCard openDelay={0} closeDelay={100}>
         <HoverCardTrigger asChild>
+          <Link href={"/dashboard"} className="flex items-center space-x-4">
           <Button variant="ghost" size="sm" className="relative size-8 rounded-full">
             <Avatar className="size-8">
               <AvatarImage src="/placeholder-user.jpg" alt="User" />
@@ -21,6 +21,7 @@ const UserButton = () => {
               </AvatarFallback>
             </Avatar>
           </Button>
+          </Link>
         </HoverCardTrigger>
         <HoverCardContent
           className="w-64 border-none bg-[#16151A] p-0 shadow-xl before:absolute before:-top-2 before:right-1 before:h-0 before:w-0 before:-translate-x-1/2 before:border-r-8 before:border-b-8 before:border-l-8 before:border-r-transparent before:border-b-[#16151A] before:border-l-transparent before:content-['']"
@@ -42,17 +43,6 @@ const UserButton = () => {
                   <p className="text-sm text-[#B3B1B0]">UID 47846465</p>
                 </div>
               </div>
-              <LoginModal TriggerButton={
-                (
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    className="text-xs"
-                  >
-                    {user ? "Log out" : "Log in"}
-                  </Button>
-                )
-              }/>
             </div>
 
             {/* Coins and Bonus */}
@@ -74,7 +64,7 @@ const UserButton = () => {
           </div>
         </HoverCardContent>
       </HoverCard>
-    </Link>
+    </div>
   );
 };
